+++ lib/visual_garden_web/live/home_live/show.html.heex
@@ -35,52 +35,3 @@
     </PC.card>
   <% end %>
 </div>
-
-<div class="prose prose-eagle mt-10">
-  <h2>Task list</h2>
-</div>
-
-<ul :for={item <- @todo_items} role="list" class="divide-y divide-gray-100">
-  <li class="flex justify-between gap-x-6 py-5">
-    <.render_todo_item
-      item={item}
-      garden={Enum.find(@gardens, &(&1.id == item.garden_id))}
-      planner_entries={@planner_entries}
-    />
-  </li>
-</ul>
-
-<.modal
-  :if={@live_action == :orphaned_nursery}
-  id="orphan-modal"
-  show
-  on_cancel={JS.patch(~p"/home")}
->
-  <.live_component
-    module={VisualGardenWeb.HomeLive.OrphanedNurseryComponent}
-    id={@entry.id}
-    nursery_entry={@entry}
-    title={@page_title}
-    current_user={@current_user}
-    patch={~p"/home"}
-  />
-</.modal>
-
-<.modal
-  :if={@live_action in [:transfer]}
-  id="event-log-modal"
-  show
-  on_cancel={JS.navigate(~p"/home")}
->
-  <.live_component
-    module={VisualGardenWeb.EventLogLive.FormComponent}
-    patch={~p"/home"}
-    action={@live_action}
-    id="event-modal"
-    product={@bed}
-    products={@products}
-    garden={@garden}
-    current_user={@current_user}
-    title={@page_title}
-  />
-</.modal>
